chore(vite-plugin): cleanup
